Extract SortTerm and SortCommandParser as outer class
